Create gtm component
