+++ apps/openchallenges/app/src/app/app.component.html
@@ -9,6 +9,7 @@
   (logInClicked)="login()"
 />
 
+<openchallenges-google-tag-manager />
 <div class="container">
   <router-outlet></router-outlet>
 </div>
