+++ apps/openchallenges/app/src/app/app.component.ts
@@ -10,16 +10,16 @@ import {
   NavbarComponent,
 } from '@sagebionetworks/openchallenges/ui';
 import { APP_SECTIONS } from './app-sections';
-import { NavigationEnd, Router, RouterOutlet } from '@angular/router';
-import { GoogleTagManagerService } from 'angular-google-tag-manager';
+import { RouterOutlet } from '@angular/router';
 import { HomeDataService } from '@sagebionetworks/openchallenges/home';
+import { GoogleTagManagerComponent } from './google-tag-manager.component';
 
 @Component({
   selector: 'openchallenges-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
   standalone: true,
-  imports: [NavbarComponent, RouterOutlet],
+  imports: [NavbarComponent, RouterOutlet, GoogleTagManagerComponent],
 })
 export class AppComponent implements OnInit, OnDestroy {
   title = 'OpenChallenges';
@@ -31,24 +31,11 @@ export class AppComponent implements OnInit, OnDestroy {
   private subscriptions: Subscription[] = [];
 
   constructor(
-    private router: Router,
     private pageTitleService: PageTitleService,
-    private gtmService: GoogleTagManagerService,
     private homeDataService: HomeDataService
   ) {}
 
   ngOnInit() {
-    this.router.events.forEach((event) => {
-      if (event instanceof NavigationEnd) {
-        const gtmTag = {
-          event: 'page',
-          pageName: event.url,
-        };
-
-        this.gtmService.pushTag(gtmTag);
-      }
-    });
-
     // TODO Call getUserProfile() only if the user is logged in, other wise an error is generated
     // when the page is rendered with SSR.
     // https://github.com/Sage-Bionetworks/sage-monorepo/issues/880#issuecomment-1318955348
