+++ libs/openchallenges/home/src/lib/home.component.html
@@ -1,5 +1,5 @@
 <main class="base mat-typography">
-  <div id="announcement">
+  <!-- <div id="announcement">
     <h4>Welcome to OpenChallenges!</h4>
     <h3>This website uses a preview snapshot of the OpenChallenges (OC) Database</h3>
     <h4>
@@ -21,7 +21,7 @@
     *sageAppShellNoRender
   />
   <openchallenges-challenge-host-list *sageAppShellNoRender/>
-  <openchallenges-sponsor-list *sageAppShellNoRender/>
+  <openchallenges-sponsor-list *sageAppShellNoRender/> -->
 </main>
 
 <openchallenges-footer
